Added enhanced sql-alchemy `NextBase`
